# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.CallRequest:
    properties:
      device_id:
        type: string
      resident_id:
        type: string
    required:
    - device_id
    - resident_id
    type: object
...
  controllers.DeviceBuildingRequest:
    properties:
      building_id:
        example: 1
        type: integer
    required:
    - building_id
    type: object
  controllers.DeviceHealthRequest:
    properties:
      device_id:
        example: "1"
        type: string
    required:
    - device_id
    type: object
...
  controllers.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      data: {}
      message:
        example: Invalid username or password
        type: string
    type: object
...
  controllers.LoginData:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      role:
        example: admin
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      password:
        example: admin123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  controllers.LoginResponse:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      message:
        example: Login successful
        type: string
    type: object
...
  controllers.TokenRequest:
    properties:
      channel_id:
        example: room123
        type: string
      user_id:
        example: user456
        type: string
    required:
    - channel_id
    - user_id
    type: object
...
host: 39.108.49.167:20033
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: A comprehensive door access management system with video calling capabilities
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: intercom_http_service API
  version: "1.0"
paths:
...
  /api/rtc/call:
    post:
      consumes:
      - application/json
      description: Initiate a video call between a device and a resident
      parameters:
      - description: Call request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Start Video Call
      tags:
      - RTC
  /api/rtc/token:
    post:
      consumes:
      - application/json
      description: Get a RTC token for real-time communication
      parameters:
      - description: Token request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get RTC Token
      tags:
      - RTC
  /auth/login:
    post:
      consumes:
      - application/json
      description: Process user login and return JWT token with different permissions
        based on user role
      parameters:
      - description: Login request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response with token
          schema:
            allOf:
            - $ref: '#/definitions/controllers.LoginResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.LoginData'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: User Login
      tags:
      - Auth
  /buildings:
    get:
      consumes:
      - application/json
      description: 获取系统中所有楼号的列表
      parameters:
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页条数，默认为10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取所有楼号
      tags:
      - Building
    post:
      consumes:
      - application/json
      description: 创建一个新的楼号
      parameters:
      - description: 楼号信息
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/controllers.BuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建楼号
      tags:
      - Building
  /buildings/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的楼号
      parameters:
      - description: 楼号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除楼号
      tags:
      - Building
    get:
      consumes:
      - application/json
      description: 根据ID获取楼号详细信息
      parameters:
      - description: 楼号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取楼号详情
      tags:
      - Building
    put:
      consumes:
      - application/json
      description: 更新楼号信息
      parameters:
      - description: 楼号ID
        in: path
        name: id
        required: true
        type: integer
      - description: 楼号信息
        in: body
        name: building
        required: true
        schema:
          $ref: '#/definitions/controllers.BuildingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新楼号
      tags:
      - Building
  /buildings/{id}/devices:
    get:
      consumes:
      - application/json
      description: 获取指定楼号关联的所有设备
      parameters:
      - description: 楼号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取楼号关联的设备
      tags:
      - Building
  /buildings/{id}/households:
    get:
      consumes:
      - application/json
      description: 获取指定楼号下的所有户号
      parameters:
      - description: 楼号ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取楼号下的户号
      tags:
      - Building
...
  /device/status:
    post:
      consumes:
      - application/json
      description: 设备用于报告在线状态的简单健康检测接口
      parameters:
      - description: 设备健康检测请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceHealthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设备健康检测
      tags:
      - device
  /devices:
    get:
      consumes:
      - application/json
      description: 获取所有设备的列表，支持按楼号筛选
      parameters:
      - description: 楼号ID
        in: query
        name: building_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取所有设备
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 创建一个新的门禁设备，支持设备类型和关联
      parameters:
      - description: 设备信息：包含名称、类型、位置、关联楼号/户号等
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceRequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: 成功创建的设备信息
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: 请求参数错误，如缺少必填字段或格式不正确
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库操作失败等
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建新设备
      tags:
      - device
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除设备
      tags:
      - device
    get:
      consumes:
      - application/json
      description: 根据ID获取设备信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取单个设备
      tags:
      - device
    put:
      consumes:
      - application/json
      description: 根据ID更新设备信息，支持更新关联
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: string
      - description: 设备信息：包含需要更新的字段
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的设备信息
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: 设备不存在
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新设备信息
      tags:
      - device
  /devices/{id}/building:
    post:
      consumes:
      - application/json
      description: 将指定设备关联到楼号
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      - description: 楼号信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceBuildingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 关联设备与楼号
      tags:
      - device
  /devices/{id}/households:
    delete:
      consumes:
      - application/json
      description: 解除指定设备与其当前关联的户号的关联
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 解除设备与户号的关联
      tags:
      - device
    get:
      consumes:
      - application/json
      description: 获取指定设备关联的所有户号
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取设备关联的户号
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 将指定设备关联到户号
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      - description: 户号信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.DeviceHouseholdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 关联设备与户号
      tags:
      - device
  /devices/{id}/status:
    get:
      consumes:
      - application/json
      description: 获取设备的当前状态信息，包括在线状态、最后更新时间等
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设备状态信息，包含ID、名称、状态、位置、最后在线时间等
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 设备不存在
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: 服务器内部错误，可能是数据库查询失败等
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取设备状态
      tags:
      - device
  /emergency/alarm:
    post:
      consumes:
      - application/json
      description: Trigger an emergency alarm and notify relevant personnel
      parameters:
      - description: Alarm request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyAlarmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trigger Emergency Alarm
      tags:
      - Emergency
  /emergency/contacts:
    get:
      consumes:
      - application/json
      description: Get a list of all emergency contacts in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Emergency Contacts
      tags:
      - Emergency
  /emergency/notify-all:
    post:
      consumes:
      - application/json
      description: Send notification to all users during an emergency
      parameters:
      - description: Notification request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.EmergencyNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Notify All Users
      tags:
      - Emergency
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"